+++ jkube-kit/enricher/generic/src/test/java/org/eclipse/jkube/enricher/generic/DefaultControllerEnricherTest.java
@@ -13,18 +13,22 @@
  */
 package org.eclipse.jkube.enricher.generic;
 
+import io.fabric8.kubernetes.api.model.Container;
 import io.fabric8.kubernetes.api.model.KubernetesList;
 import io.fabric8.kubernetes.api.model.KubernetesListBuilder;
+import io.fabric8.kubernetes.api.model.PodSpec;
+import io.fabric8.kubernetes.api.model.PodTemplateSpec;
+import io.fabric8.kubernetes.api.model.apps.Deployment;
+import io.fabric8.kubernetes.api.model.apps.DeploymentSpec;
+import org.assertj.core.api.InstanceOfAssertFactories;
 import org.eclipse.jkube.kit.common.KitLogger;
 import org.eclipse.jkube.kit.config.image.build.BuildConfiguration;
 import org.eclipse.jkube.kit.config.image.ImageConfiguration;
 import org.eclipse.jkube.kit.common.JavaProject;
 import org.eclipse.jkube.kit.config.resource.PlatformMode;
 import org.eclipse.jkube.kit.config.resource.ProcessorConfig;
-import org.eclipse.jkube.kit.config.resource.ResourceConfig;
 import org.eclipse.jkube.kit.enricher.api.EnricherContext;
 import org.eclipse.jkube.kit.enricher.api.JKubeEnricherContext;
-import org.eclipse.jkube.kit.common.util.ResourceUtil;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -43,21 +47,22 @@ public class DefaultControllerEnricherTest {
 
     private Map<String, Map<String, Object>> config;
     private EnricherContext context;
+    private Properties properties;
 
     @Before
     public void setUp() throws Exception {
         config = new HashMap<>();
+        properties = new Properties();
         context = JKubeEnricherContext.builder()
             .processorConfig(new ProcessorConfig(null, null, config))
             .log(new KitLogger.SilentLogger())
-            .resources(ResourceConfig.builder().build())
             .image(ImageConfiguration.builder()
                 .name("helloworld")
                 .build(BuildConfiguration.builder()
                     .port("8080")
                     .build()).build())
             .project(JavaProject.builder()
-                .properties(new Properties())
+                .properties(properties)
                 .groupId("group")
                 .artifactId("artifact-id")
                 .build())
@@ -65,20 +70,45 @@ public class DefaultControllerEnricherTest {
     }
 
     @Test
-    public void checkReplicaCount() throws Exception {
-        enrichAndAssert(3);
+    public void checkReplicaCount() {
+        givenReplicaCountInEnricherConfig(String.valueOf(3));
+        enrichAndAssert(3, "IfNotPresent");
     }
 
     @Test
-    public void checkDefaultReplicaCount() throws Exception {
-        enrichAndAssert(1);
+    public void checkDefaultReplicaCount() {
+        givenReplicaCountInEnricherConfig(String.valueOf(1));
+        enrichAndAssert(1, "IfNotPresent");
     }
 
-    protected void enrichAndAssert(int replicaCount) throws Exception {
-        final Map<String, Object> controllerConfig = new TreeMap<>();
-        controllerConfig.put("replicaCount", String.valueOf(replicaCount));
-        config.put("jkube-controller", controllerConfig);
+    @Test
+    public void create_withImagePullPolicyConfigured_shouldUseConfiguredImagePullPolicy() {
+        // Given
+        properties.put("jkube.enricher.jkube-controller.pullPolicy", "Never");
 
+        // When + Then
+        enrichAndAssert(1, "Never");
+    }
+
+    @Test
+    public void create_withImagePullPolicyProperty_shouldUseConfiguredImagePullPolicy() {
+        // Given
+        properties.put("jkube.imagePullPolicy", "Never");
+
+        // When + Then
+        enrichAndAssert(1, "Never");
+    }
+
+    @Test
+    public void create_withReplicasConfigured_shouldUseConfiguredReplicas() {
+        // Given
+        properties.put("jkube.enricher.jkube-controller.replicaCount", "5");
+
+        // When + Then
+        enrichAndAssert(5, "IfNotPresent");
+    }
+
+    protected void enrichAndAssert(int replicaCount, String imagePullPolicy) {
         // Enrich
         DefaultControllerEnricher controllerEnricher = new DefaultControllerEnricher(context);
         KubernetesListBuilder builder = new KubernetesListBuilder();
@@ -88,11 +118,25 @@ public class DefaultControllerEnricherTest {
         KubernetesList list = builder.build();
         assertEquals(1, list.getItems().size());
 
-        String json = ResourceUtil.toJson(list.getItems().get(0));
         assertThat(list.getItems())
-                .hasSize(1)
-                .first()
-                .hasFieldOrPropertyWithValue("spec.replicas", replicaCount);
+            .hasSize(1)
+            .first()
+            .asInstanceOf(InstanceOfAssertFactories.type(Deployment.class))
+            .hasFieldOrPropertyWithValue("metadata.name", "artifact-id")
+            .hasFieldOrPropertyWithValue("spec.replicas", replicaCount)
+            .extracting(Deployment::getSpec)
+            .extracting(DeploymentSpec::getTemplate)
+            .extracting(PodTemplateSpec::getSpec)
+            .extracting(PodSpec::getContainers)
+            .asList()
+            .first(InstanceOfAssertFactories.type(Container.class))
+            .hasFieldOrPropertyWithValue("imagePullPolicy", imagePullPolicy);
+    }
+
+    private void givenReplicaCountInEnricherConfig(String value) {
+        final Map<String, Object> controllerConfig = new TreeMap<>();
+        controllerConfig.put("replicaCount", value);
+        config.put("jkube-controller", controllerConfig);
     }
 
 }