+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Usage:
 * Fix #1468: Add startup probe in QuarkusHealthCheckEnricher
 * Fix #1473: Add OpenLibertyHealthCheckEnricher to automatically add health checks in OpenLiberty projects
 * Fix #1537: Registry not set up in `oc:build`
+* Fix #1619: property `jkube.enricher.jkube-controller.pullPolicy` is not taken into account if a `deployment` resource fragment is defined
 * Fix #1654: images with missing build configuration should not be built